FB-15a: extract buildApiRequestParams into TaskRequestBuilder.ts
